Adjusted CSS
Adjusted the css for the elements within the html file according to the README.md note while comparing what was in the lesson and testing where to place the word light to make the background color ACTUALLY light green. But, after double checking the lesson and the README.md file, I was able to complete the assignment by double checking it with the desirable outcome image.