Readjusted CSS
Fixed some the css to accomodate alternative styling for the text to fit more in line with the solution after looking at it. If you were to compare the 2, you would see that I didn't just copy what was there and tried to figure out how to accomplish the same task. Mine looks a bit more sloppy to me, but hey, it's a start.